requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.6.10
django-classy-tags==0.6.1
django-cms==3.0.9
django-mptt==0.6.1
django-reversion==1.8.5
django-sekizai==0.8.1
djangocms-admin-style==0.2.5
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.7.2
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-video==0.1
djangorestframework==3.0.5
djangorestframework-xml==1.0.1
html5lib==0.999
mysqlclient==1.3.6
Pillow==2.7.0
psycopg2==2.6
pytz==2014.10
requests==2.5.1
six==1.9.0
South==1.0.2
